feat:增加逻辑删除功能，防止误删

+++ t3rik-hrm/src/main/java/com/t3rik/hrm/sm/domain/HrmStaff.java
@@ -136,7 +136,7 @@ public class HrmStaff extends BaseEntity
      * 逻辑删除字段 0:未删除 1:已删除
      */
     @TableLogic
-    private Long deleted;
+    private Integer deleted;
 
     /**
      * 逻辑删除辅助字段
